:sparkles: 스토리보드에서 ContainerView 없애고 UI 다시 구현
